+++ .markdownlint.mjs
@@ -6,7 +6,7 @@ SPDX-License-Identifier: CC-BY-4.0 OR GPL-3.0-or-later
 This file is part of Network Pro.
 ========================================================================== */
 
-import noSmartQuotes from './.md-smart-quotes.js'; // adjust path as needed
+import noSmartQuotes from './.md-smart-quotes.js';
 
 export default {
   // Enable all default rules, then override below